Organized files ... comments throughout code will need to be made.
