--- gnp/src/main/java/com/gnp/auth/AuthConstants.java
@@ -1,11 +1,0 @@
-package com.gnp.auth;
-
-public class AuthConstants {
-	  public static final String SECRET = "SuperSecretKey";
-	  public static final long EXPIRATION_TIME = 432_000_000; 
-	  public static final String TOKEN_PREFIX = "Bearer ";
-	  public static final String HEADER_STRING = "Authorization";
-	  public static final String SIGN_UP_URL = "/api/user/register";
-	  public static final String SALES_URL = "/sales";
-	  public static final String ESTIMATE_URL = "/estimate";
-	}
