--- gnp/src/main/java/com/gnp/auth/ERole.java
@@ -1,7 +1,0 @@
-package com.gnp.auth;
-
-public enum ERole {
-	ROLE_USER,
-    ROLE_MODERATOR,
-    ROLE_ADMIN
-}
