--- gnp/src/main/java/com/gnp/auth/JWTAuthorizationFilter.java
@@ -1,47 +1,0 @@
-package com.gnp.auth;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import javax.servlet.*;
-import javax.servlet.http.*;
-import org.springframework.security.authentication.*;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-
-import static com.gnp.auth.AuthConstants.*;
-
-public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
-  public JWTAuthorizationFilter(AuthenticationManager authenticationManager) {
-    super(authenticationManager);
-  }
-  
-  @Override
-  protected void doFilterInternal(HttpServletRequest req, HttpServletResponse res, FilterChain chain) throws IOException, ServletException {
-    String header = req.getHeader(HEADER_STRING);
-    if (header == null || !header.startsWith(TOKEN_PREFIX)) {
-      chain.doFilter(req, res);
-      return;
-    }
-    UsernamePasswordAuthenticationToken authentication = getAuthentication(req);
-    SecurityContextHolder.getContext().setAuthentication(authentication);
-    chain.doFilter(req, res);
-  }
-  
-  private UsernamePasswordAuthenticationToken getAuthentication(HttpServletRequest request) {
-    String token = request.getHeader(HEADER_STRING);
-    if (token != null) {
-      String user = JWT.require(Algorithm.HMAC512(SECRET.getBytes()))
-      .build()
-      .verify(token.replace(TOKEN_PREFIX, ""))
-      .getSubject();
-      
-  	  if (user != null) {
-        return new UsernamePasswordAuthenticationToken(user, null, new ArrayList<>());
-      }
-      return null;
-    }
-    return null;
-  }
-}
