+++ gnp/src/main/java/jwt/AuthEntryPointJwt.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package jwt;
 
 import java.io.IOException;
