+++ gnp/src/main/java/models/ERole.java
@@ -1,0 +1,7 @@
+package models;
+
+public enum ERole {
+	ROLE_USER,
+	ROLE_MODERATOR,
+	ROLE_ADMIN
+}
