+++ gnp/src/main/java/repos/RoleRepository.java
@@ -1,9 +1,12 @@
-package com.gnp.auth;
+package repos;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import models.ERole;
+import models.Role;
 
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
