+++ gnp/src/main/java/repos/SalesRepository.java
@@ -1,8 +1,10 @@
-package com.gnp.auth;
+package repos;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import models.Sales;
 
 public interface SalesRepository extends JpaRepository<Sales, Integer> {
 	List<Sales> findFirst28ByOrderByDateDesc();
