+++ gnp/src/main/java/repos/UserRepository.java
@@ -1,13 +1,18 @@
-package com.gnp.auth;
+package repos;
+
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import models.User;
+
+//This is called in AuthController
+
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByUsername(String username);
-  
-  Boolean existsByUsername(String username);
+
+	Boolean existsByUsername(String username);
 
 	Boolean existsByEmail(String email);
 }
