+++ gnp/src/main/java/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package request;
 
 import javax.validation.constraints.NotBlank;
