+++ gnp/src/main/java/response/MessageResponse.java
@@ -1,11 +1,11 @@
-package com.gnp.auth;
+package response;
 
 public class MessageResponse {
 	private String message;
 
 	public MessageResponse(String message) {
-	    this.message = message;
-	  }
+		this.message = message;
+	}
 
 	public String getMessage() {
 		return message;
