+++ gnp/src/main/java/services/UserDetailsServiceImpl.java
@@ -1,5 +1,4 @@
-package com.gnp.auth;
-
+package services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -7,6 +6,9 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import models.User;
+import repos.UserRepository;
 
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
